+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Set up unit tests environment. [#7] (https://github.com/rokwire/core-building-block/issues/7)
+- Expand the model to handle the user devices. [#41] (https://github.com/rokwire/core-building-block/issues/41)
 - Expose get global config admin API [#35] (https://github.com/rokwire/core-building-block/pull/43)
 - Set up auth framework. [#16] (https://github.com/rokwire/core-building-block/issues/16)
 - Expose version API. [#13] (https://github.com/rokwire/core-building-block/issues/13)